@@ -5,46 +5,39 @@
   "main": "./lib/loader.js",
   "types": "./lib/loader.d.ts",
   "author": "Stephen Wan <stephen@stephenwan.com>",
-  "homepage": "https://github.com/stephen/graphql-loader/",
+  "homepage": "https://github.com/jorrit/graphql-loader/",
   "repository": {
     "type": "git",
-    "url": "https://github.com/stephen/graphql-loader.git"
+    "url": "https://github.com/jorrit/graphql-loader.git"
   },
   "license": "MIT",
   "scripts": {
     "build": "tsc",
-    "prepublish": "npm test && npm run build",
+    "prepare": "npm run build",
     "test": "jest"
   },
-  "resolutions": {
-    "graphql": "0.13.2"
-  },
+  "files": [
+    "lib"
+  ],
   "devDependencies": {
-    "@types/graphql": "^0.11.7",
-    "@types/jest": "^20.0.4",
-    "@types/loader-utils": "^1.1.0",
-    "@types/memory-fs": "^0.3.0",
-    "@types/node": "^8.0.14",
-    "@types/webpack": "^3.0.4",
-    "graphql": "0.13.2",
-    "jest": "^20.0.4",
-    "memory-fs": "^0.4.1",
-    "prettier": "^1.5.3",
-    "ts-loader": "3.x",
-    "ts-node": "^3.2.0",
-    "typescript": "2.6.1",
-    "webpack": "3.8.1"
+    "@types/jest": "^28.1.1",
+    "@types/memory-fs": "^0.3.3",
+    "@types/node": "^16.0.0",
+    "@types/webpack": "^5.0.0",
+    "graphql": "^16.5.0",
+    "jest": "^28.1.1",
+    "memory-fs": "^0.5.0",
+    "prettier": "^2.7.0",
+    "ts-jest": "^28.0.5",
+    "typescript": "^4.7.0",
+    "webpack": "^5.73.0"
   },
-  "dependencies": {
-    "apollo-codegen": "^0.19.1",
-    "loader-utils": "^1.1.0",
-    "pify": "^3.0.0"
-  },
+  "dependencies": {},
   "peerDependencies": {
-    "graphql": ">=0.11.7",
-    "ts-loader": "^4.4.1"
+    "graphql": "^16.5.0"
   },
   "jest": {
+    "preset": "ts-jest",
     "testMatch": [
       "**/test/?(*.)(spec).ts"
     ],
@@ -52,9 +45,6 @@
       "ts",
       "tsx",
       "js"
-    ],
-    "transform": {
-      "^.+\\.(ts|tsx)$": "<rootDir>/test/preprocessor.js"
-    }
+    ]
   }
 }
